+++ src/Kiota.Builder/Writers/Shell/ShellCodeMethodWriter.cs
@@ -103,7 +103,7 @@ namespace Kiota.Builder.Writers.Shell
 
             AddCustomCommandOptions(writer, ref availableOptions, ref paramTypes, ref paramNames, returnType, isHandlerVoid);
 
-            if (!isHandlerVoid && !conventions.StreamTypeName.Equals(returnType, StringComparison.OrdinalIgnoreCase))
+            if (!isHandlerVoid && !conventions.StreamTypeName.Equals(returnType, StringComparison.OrdinalIgnoreCase) && !conventions.IsPrimitiveType(returnType))
             {
                 // Add output filter param
                 paramNames.Add(outputFilterParamName);
@@ -160,7 +160,7 @@ namespace Kiota.Builder.Writers.Shell
                 paramNames.Add(fileParamName);
                 availableOptions.Add(fileOptionName);
             } 
-            else if (!isHandlerVoid)
+            else if (!isHandlerVoid && !conventions.IsPrimitiveType(returnType))
             {
                 // Add output type param
                 var outputOptionName = "outputOption";
@@ -214,9 +214,17 @@ namespace Kiota.Builder.Writers.Shell
                 if (typeString != "Stream")
                 {
                     var formatterOptionsVar = "formatterOptions";
-                    writer.WriteLine($"var {formatterVar} = {outputFormatterFactoryParamName}.GetFormatter({outputFormatParamName});");
-                    writer.WriteLine($"response = await {outputFilterParamName}?.FilterOutputAsync(response, {outputFilterQueryParamName}, {cancellationTokenParamName}) ?? response;");
-                    writer.WriteLine($"var {formatterOptionsVar} = {outputFormatParamName}.GetOutputFormatterOptions(new FormatterOptionsModel(!{jsonNoIndentParamName}));");
+                    var formatterTypeVal = "FormatterType.TEXT";
+                    if (conventions.IsPrimitiveType(typeString))
+                    {
+                        writer.WriteLine($"IOutputFormatterOptions? {formatterOptionsVar} = null;");
+                    } else
+                    {
+                        formatterTypeVal = outputFormatParamName;
+                        writer.WriteLine($"response = await {outputFilterParamName}?.FilterOutputAsync(response, {outputFilterQueryParamName}, {cancellationTokenParamName}) ?? response;");
+                        writer.WriteLine($"var {formatterOptionsVar} = {outputFormatParamName}.GetOutputFormatterOptions(new FormatterOptionsModel(!{jsonNoIndentParamName}));");
+                    }
+                    writer.WriteLine($"var {formatterVar} = {outputFormatterFactoryParamName}.GetFormatter({formatterTypeVal});");
                     writer.WriteLine($"await {formatterVar}.WriteOutputAsync(response, {formatterOptionsVar}, {cancellationTokenParamName});");
                 }
                 else
@@ -307,9 +315,8 @@ namespace Kiota.Builder.Writers.Shell
 
             if ((codeElement.AccessedProperty?.Type) is CodeType codeReturnType)
             {
-                // Include namespace to avoid type ambiguity on similarly named classes. Currently, if we have namespaces A and A.B where both namespaces have type T,
-                // Trying to use type A.B.T in namespace A without using the fully qualified name will break the build.
-                var targetClass = string.Join(".", codeReturnType.TypeDefinition.Parent.Name, conventions.GetTypeString(codeReturnType, codeElement));
+                var targetClass = conventions.GetTypeString(codeReturnType, codeElement);
+
                 var builderMethods = codeReturnType.TypeDefinition.GetChildElements(true).OfType<CodeMethod>()
                     .Where(m => m.IsOfKind(CodeMethodKind.CommandBuilder))
                     .OrderBy(m => m.Name)