+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added text output formatter to CLI commons. (Shell)
 - Added support for vendor specific content types generation/serialization. [#1197](https://github.com/microsoft/kiota/issues/1197)
 - Added support for 204 no content in generation and CSharp/Java/Go/TypeScript request adapters. #1410
 - Added a draft swift generation implementation. #1444
@@ -16,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fixed a bug where request builder classes for collections endpoints would have a wrong name. #1052
+- Fix issue with ambiguous type names causing build errors and stack overflows. (Shell) #1052
 - Fixed a bug where symbols (properties, methods, classes) could contain invalid characters #1436
 
 ## [0.0.20] - 2022-03-25